add generic base menu for navbar menus
