+++ src/components/navigation/navbar/student/CoursesMenu/CourseMenu.vue
@@ -1,41 +1,25 @@
 <template>
-    <div class="relative cursor-pointer" @mouseleave="hide" @mouseover="show">
-        <div>Courses</div>
-
-        <transition name="menu-fade">
-            <div
-                v-show="isVisible"
-                class="absolute normal-case font-normal bg-white shadow-md rounded-lg overflow-hidden border mt-4 z-30 lg:z-10 left-0 lg:w-180 lg:-left-16"
-            >
+    <base-menu>
+        <template #hook>Courses</template>
+        <template #content>
+            <div class="absolute left-0 lg:-left-16 lg:w-160 mt-4 z-30 lg:z-10 bg-white shadow-md rounded-lg overflow-hidden border">
                 <menu-body />
                 <menu-footer />
             </div>
-        </transition>
-    </div>
+        </template>
+    </base-menu>
 </template>
 
-<script>
-    import MenuBody from "./CourseMenuBody";
-    import MenuFooter from "./CourseMenuFooter";
-    import { ref } from "vue";
+<script lang="ts">
+    import BaseMenu from "@/components/navigation/navbar/common/BaseMenu.vue";
+    import MenuBody from "./CourseMenuBody.vue";
+    import MenuFooter from "./CourseMenuFooter.vue";
 
     export default {
         components: {
+            BaseMenu,
             MenuBody,
             MenuFooter,
-        },
-        setup() {
-            const isVisible = ref(false);
-
-            function show() {
-                isVisible.value = true;
-            }
-
-            function hide() {
-                isVisible.value = false;
-            }
-
-            return { isVisible, show, hide };
         },
     };
 </script>
